Add files via upload
This one is a personal project. Having learned how to use the yfinance library in Python, I thought I would use the skill to collect price data from all the 14 assets in my portfolio, and joined them using Posgresql to make all sorts of analysis on them. I made two distinct joins, an inner join for my studies requiring balanced samples (correlation scatterplot and volatility) and a full outer to observe price growth and cumulative % growth of the portfolio (as well as each individual asset) across time. I am still working on it, but I've had a lot of fun with it so far.